Remove obsolete test_logging.py and add comprehensive logging tests in test_core_logging.py, including system log handling and unified logging functionality. Update FakeVariable in test_core_block_helpers.py to enable type checking.
